refactor(credit): modularize portfolio_intelligence → portfolio/ package (Wave 1 PR #8)
Split 797-LOC monolithic portfolio_intelligence.py into edgar-style package:
- models.py: PORTFOLIO_CONTAINER constant, safe_float/extract_percent helpers
- discovery.py: discover_active_investments
- metrics.py: extract_portfolio_metrics
- drift.py: detect_performance_drift + _latest_metric_by_investment
- covenants.py: build_covenant_surveillance
- risk.py: reclassify_investment_risk + _evaluate_liquidity_cash_impact
- briefs.py: build_board_monitoring_briefs
- service.py: run_portfolio_ingest orchestrator
- __init__.py: standard imports (no PEP 562 — stdlib only)

Migrated logging → structlog. Never-raises error contract.
Added PortfolioMetric stub (pre-existing missing model).
Updated 3 callers: ai_engine/__init__.py, ai/portfolio.py, test_vertical_engines.py.
