+++ docs/plans/2026-03-15-refactor-credit-vertical-modular-alignment-wave1-plan.md
@@ -333,7 +333,7 @@ market_data/
 
 ---
 
-### Phase 8: portfolio/ (PR #8)
+### Phase 8: portfolio/ (PR #8) ✅
 
 **Source:** `portfolio_intelligence.py` (797 LOC)
